Adjusted things to remove jerkiness from the intro text.
Because the intro text is drawn line-by-line, advancing by a floating
point iteration for each line caused rounding to offset odd lines from
even lines in a way that would switch over time, creating a horribly
jerky effect. I assume this is probably GPU dependent, but at least
on my system, it looked horrible. This changes it so that the line
height used for printing line by line is always and integer: the
calculated line height is ceil()ified to ensure this. Makes it look
much, much better.