Replaced the extreneous "1" notation of boolean stats with just a tag.
This makes it so that boolean stats (instant jump, reverse thrust,
and asteroid scan) no longer require you to specify a "1" in XML.
You just put the tag and the engine automatically assumes you're
enabling that stat. The reason for this is pretty simple: a zero
value is meaningless because it always does absolutely nothing.
Stats were being added together and then converted back into boolean,
so if you had a zero value alongside a one value, it would become
one every time.

Also changed that method from addition to or, since that's what it
actually means.