Added Reload/Rebuild Time display, improved launcher stats display a bit.
I noticed that the current launchers didn't have enough room for
another line, so I made an optimization of putting the specification
of seeker / unguided right next to the type name. This has the added
benefit of also decoupling this from lock-on time, which is not
actually required for seekers anymore.