Made it so that the Brushed GUI shows instant mode weapons.
This is something that I had to make an API change for, and I must
admit my implementation is a bit inefficient and somewhat messy. In
any case, what this does is add a new "instant" parameter. This was
necessary because otherwise the GUI could only either show every
weapon on the ship, or only those in the selected weapon set (which
excludes instant mode weapons).