Expanded the Ship Stealing mission to allow stealing non-designated ships.
This causes pirate informers to offer to help steal ships other than
your target when you board them… for an extra fee. The fee is
calculated based on the difference of what you paid and what the
boarded ship is worth. If you already paid more than the ship's
value, there is no extra charge. What this means is that stealing
non-designated targets is never profitable, but can be used to steal
a ship you really want to add to your collection.