Lower ammo of steal targets… which required some new Lua API.
Wow, what a rabbit hole. Turned out there was nothing in the API
for handling ammo beyond adding it, so I added two new functions
and extended pilot.outfitRm() to handle removing ammo as well. (One
of the two new functions turned out to be unnecessary for what I
wanted to do, but no harm in keeping it.)