Added support for disabling radar, use it while overlay open in Brushed/slim.
This turned out to be slightly more complicated than I expected, but
oh well. This feature (making the radar disappear and stop working)
is very useful because the radar can easily get in the way of trying
to click on overlay things, or make it hard to tell what location the
indicator dots are for, particularly for pilots.

However, this isn't ideal for all GUIs (for example it's completely
unnecessary in Legacy), so I made this something that is handled on
the Lua side by the GUI. This also has the advantage of enabling
checking if the overlay is open or disabling the radar for other
interface reasons, if applicable (one example that comes to mind is
it would enable creating a fancy graphical border around the overlay,
or getting rid of a graphical border used for the radar).