Fixed the ancient fallback for slot size color in slot rendering.
I trawled through Naev's old history, and boy, is the logic there
ancient. It turns out that it was always intended as a fallback for
undefined slot sizes, and the comment wasn't in reference to what
was left, but very ancient logic that checked slot size inline
manually (with copy-pasted code in two separate places), but over
the years the method got edited and the fallback accidentally ended
up only working on the selected case. Deiz saw that cut down code,
saw that it was doing very little, and trimmed it down to its actual
effects, not noticing that the selected check was in fact a leftover
from before then, and should have been removed.

Realistically, this doesn't change anything since it was actually
a fail condition (it can only possibly happen if the slot size is
something invalid). In light of that, I've added a warning alongside
the fallback coloring. I've also changed the fallback coloring to
red so that the presence of an error is more obvious.