Proper fix to the DIY-Nerds segfault: rearrange landing steps.
This avoids that horrible hack of a fix and instead fixes the root
of the problem, which was that hooks were being run before the whole
land window was even finished generating. It now runs hooks after
this. I genuinely don't know why it was doing it the other way, but
maybe further testing will reveal some unforeseen problem. Will have
to keep my eyes peeled.