Take 3: truly proper fix.
The last attempt did have a side effect as I feared: it made missions
not show up.

So, here's a completely different solution, basically the same as
the old one except with a check to see if already generated, and also
duplicated for all other tabs. This seems to put the bug to rest
once and for all with no side effects that I can see.