Fixed Baron Prowl happening alongside Baron missions.
Some time ago I decided to change its method to claims, but forgot
to actually insert the claim. Oops! This fixes that, and also uses
the resulting code change to simplify a bit (don't need to constantly
make API calls to get the planet when we've got it cached in a
variable).