Fixed a long-standing AI bug preventing pilot.land() from working.
The bug caused pilot.land() with no second argument to fail because
the code that chose a planet failed to supply a position for the
mem.land parameter, and instead provided a planet, which was invalid.
So far as I can tell, this was broken for years and just never came
up because no code was using pilot.land() in this way.