Fixed a crash when starting Naikari in a nebula system, adjusted volat warning.
Hey, a bug that affects only 0.2.0 for once!

This was caused by code that assumed that player.p existed, which isn't
true when you're in the main menu. Not sure how I didn't notice this
problem sooner.

While I fixed this, I also reformatted the volatility damage warning.
This should be easier to read.