Removed the Time Constant tutorial.
This was originally added to inform the player of how to equip the
ship and because someone found the system confusing, but default
equipment now makes how to equip ships self-documenting and I think
it's not really helpful to draw attention to it otherwise as it
gives a false impression that the feature is more important for
combat than it actually is. It makes me wonder if this tutorial is
why multiple people are scapegoating the system to blame on all of
Naev's balance problems.

The Time Constant values are actually significantly lower than they
were when the tutorial was added, as well, so most people should
find it to be unobtrusive. The fact that it's called "Time Constant"
instead of the confusing "Time Dilation" bobbens originally approved
in Naev also makes it more obvious what that value is.