Completely removed the "log type" categorization system.
"Log type" overcomplicated things, was almost never used, and even
if you did use it, it didn't help the player out very much (and in
fact just kind of made things harder to find).

This currently leaves a gaping hole in the log interface. It's only
a cosmetic defect and I'm out of spoons, so I'll fix it later. Likely
what I'll also do at some point is consolidate some of the logs
together (already did this with tutorial logs).