Adjusted the ship log; now shows entries on the right of the window.
The "View Log" button is still there because there currently isn't
enough space to see really long logs (only one that I noticed,
currently). We might want to consider making the window bigger, but
that's a more involved project which requires reorganizing some
things that are designed to work on a small window.