Removed the "Welcome to Naikari" message.
Some background: the "Welcome to Naikari" message was just a slightly
modified variant of Naev's "Welcome to Naev" message, which was
something we added to Naev at some point because Naev is incorrectly
marked as a complete game on Steam. Naikari doesn't have that problem;
on all platforms it is clearly indicated as early access. The
welcome message is therefore unnecessary.

It is also something that should be removed because Naikari is being
updated more frequently than Naev's "maybe once a year" pattern. It
could get annoying after awhile to see the same message every month.