Fixed a bug in calculation of distances for pirate cargo missions.
Code that was checking distances wasn't taking into account
hypergates. Fixed this by adding an additional API feature into
the getsysatdistance() function and also making it so that cargo
missions pass on the jump distance to the cargo description function,
rather that distance being calculated there (so the correct distance
is reported).

Seems to work, though it's probably best to hold this back from the
0.2.0 release because of the API changes. We checked thoroughly and
there shouldn't be any breakages, but better safe than sorry since
the change fixes a relatively minor bug and a regression would likely
be worse if it did happen.