Made it so that ships are never given 0 tonnes of cargo by equip script.
This fixes a cosmetic issue where you could end up with 0 tonnes of
cargo from looting. I'm honestly not sure why it was designed to add
0 tonnes of standard cargo to ships.

This also changes the method: previously, the probability of cargo
was also used as a "maximum cargo" setting, which seems very strange
to me and meant that a full NPC pilot cargo hull was impossible.