Restructured the FLF campaign so the hero and dvk parts are separate.
This begins the adoption of a different paradigm that is more
"decentralized": campaign missions slowly bump the cap, rather than
setting the cap to a specific value, more similar to the Empire
campaign. The DVK campaign will become the "main" campaign, while the
Hero campaign will be a "side" campaign that helps to build your
reputation. I may add other "side" campaigns as well. Hope is that
this new structure will make the FLF campaign feel less linear and
more dynamic.

This does temporarily cut off the "Diversion from Raelid" mission.
I'll try to fix that very soon with some new FLF Hero missions which
will bump the cap up to 30 when completed.