Reworked the Trader Escort mission so that it uses system claims.
This rework causes the trader escort mission to be higher priority,
and lowers its chance of appearing a bit. A side effect of the change
is that only one escort mission can ever be active at a time, which
is probably for the best.

This is necesary because not making system claims allows other missions
to wreak havok with it e.g. by clearing the system.