Swapped dvk01 and hero01.
This changes each to be in the other's campaign, and also changes the
mission-giving character accordingly. The reason for this is the rescue
mission is more similar to the "hero" campaign's focus while the
opportunistic destruction mission is more similar to the "Dvaered kill"
campaign's focus.

This also adds a justification for the wonton destruction in what is now
dvk01: both warlords are gung-ho on invading the Frontier, so taking them
out should temporarily weaken a desire to invade the Frontier.