Removed thrust from Mace Rockets.
I had no idea just how horrendously broken the thrust mechanic is.
Routinely, the calculation for range overestimates by tremendous
amounts, sometimes thinking that there's more than twice as much
range as there is. It's clear that we need a rehaul and that thrust
cannot ever work properly on mace rockets as they are now. Probably
what I'll do is keep the thrust mechanic only for homing missiles
and have the initial speed be the max speed, using thrust for
corrections only.