Corrected the calculation for beam heat-up.
For some reason the calculation for beam heat was based on the
calculation for bolt heat with odd extras tacked-on. This doesn't
make sense because bullets generate one-time *batches* of heat.
Beams, on the other hand, work more like afterburners and generate
a steady supply of heat.

Also made the formulas governing these "heat" amounts easier to read.
Some were written in such a way that it looked like it did something
different than what it actually did. Also, got rid of those gross
trailing operators.