Added a Money display to the equipment screen.
This required squishing some of the existing info down, so I put the
class on the same line as the ship model. I considered doing the same
on the shipyard and info screens for consistency, but the info screen
isn't wanting for space and the shipyard is a place where the class
needs to be easily visible (it's less important on the Equipment screen).