Added a Net Worth stat to the info window.
Just a cosmetic thing, and a nice little way to track progress,
kind of like a score. It doesn't count commodities since that would
complicate things (it's not obvious what price should be used), and
it's not likely to make much of a difference to the net worth figure
in general.