Adjusted race2 and prince missions.
This was supposed to be two commits, but I forgot to commit the race2
change first, so I suppose I'll just allow them to be a single commit.

The adjustment for race2 reduces the system radius used by it to
half the system's radius, rather than the full system radius. This
reduces the chances of one of the checkpoints spawning outside of the
visible area considerably, though in theory doesn't necessarily eliminate
this problem.

For the Prince mission, the only adjustment is to the instructions given
for how to board the Pinnacle, simplifying them under the assumption
that the player probably doesn't need such an in-depth reminder of the
procedure after doing it the first time in the Baron intro mission.