Experimental: prioritize wayland driver over x11 driver.
This requires a feature newly added to SDL 2.0.22, and it has to be
compiled with that version to work (since SDL_HINT_VIDEODRIVER is
a new CPP macro). However, this may hopefully eliminate the
long-standing problem where the game would crash under Wayland for
many systems, at least where SDL 2.0.22 or later is available.
To retain compatibility with older SDL versions, this is under an
ifdef CPP check (which also means it will usually do nothing).