Added a couple new requirements for the FLF Derelict event.
The new requirements are: that the current system must not have
Frontier presence, and that the player must have a Mercenary License.
The disallowing of Frontier presence is to avoid a weird situation
where the FLF pilot wants you to take him away from Frontier space.
The Mercenary License requirement is simply to ensure that players
who aren't ready for combat yet don't get entered into either the
FLF campaign or the Dvaered campaign.