Reorganized the constants used for menus to guard against future mistakes.
The constants used were required to align with other attributes of
menus and graphical bugs would be caused if something was changed but
the constants weren't updated. This changes things so that height for
menus is determined with a formula (still needs manual intervention but
is next to the place where changes that cause such a need would be)
and makes all menus use the same window width constant, since they all
also use the same button width (which is designed to correspond).