Changed Info window to Ship Computer, also corrected gui.menuInfo.
"Info Window" is a rather odd term given what the window actually
does, serving as a central hub for everything pertaining to your
ship and journey. "Ship Computer" is a more apt and clear name for
it.

While I was doing the conversion, I noticed a couple small things
and fixed them: the usage of two mission steps rather than just one
in the Waste Dump mission, and the gui.menuInfo function not
supporting the ship log as a target.