Improved the load game menu a bit and started work on snapshots.
I was hoping to finish snapshots today but didn't quite get far
enough, so the code added for that purpose is still not functional.
That said, this improves the load menu by enlarging the window and
the info text on the right, as well as making better use of space.