Started converting the tutorial to something better.
This changes the context of the mission from being a salesperson to
being an instructor employed by Melendez teaching you how to pilot
your ship. I've also added a small reward for completing the mission
to hopefully make it feel more worthwhile, and removed the prompt
asking if you want to do the tutorial. Hopefully this should make the
tutorial better. Will be doing more work on this hopefully tomorrow.