Removed most enter tutorial events and reconfigured the remaining one.
The events explaining nebula volatility and what to do about pirates
were a bit excessive, so they've been removed. As for the tutorial
on refueling from other pilots, I changed the context to another pilot
being in the same situation as you and asking for fuel, which should
serve as a more natural introduction to the mechanic.