Substantially improved the trader escort mission.
This causes the traders to actually make a formation rather than
just speed-matching as before. Seems to work well, though unfortunately
the groups are so good at surviving on their own that I wasn't able
to test the code that handles pilot death. Will have to do more
testing later (and also evaluate whether the reward for this mission
is too high).