Converted most time units to being named after real ones.
This almost matches a proposal I made years ago which the community
largely rejected: basing the time names on real time names. I'm
sympathetic to wanting the time system to feel more mystical and
disconnected from ours, but there are two problems: it's potentially
confusing (adding an extra layer of complexity to the lore that
requires extra teaching), and it doesn't really make much sense. I
just can't see all humans everywhere abandoning use of the time
units we're used to based on Earth's rotation in favor of confusing
generic names like "periods". "Cycles" is believable, changing the
meaning of existing time units seems reasonable (especially since
the new definitions are so close to the old ones), but the idea of
people confusingly using the word "period" everywhere while also
continuing to use that word for its real-world meaning is kind of
silly.