Added tooltips over the shipyard slot squares.
This has been a problem I've known about for a long time: the slots
display of a bunch of squares means nothing to new players. When I
saw someone playing Naev also pointing that out, I finally decided
to leap into action on this, and the solution I've come up with is
to add tooltips when hovering the mouse over these squares. This
makes the display much more accessible while still allowing the
quick glance thing for experienced players.