Removed the "shipyard can swap" distinction.
Previously, you needed to be on a planet with a shipyard, specifically,
to do things with your other ships. This distinction made sense in the
lore, but was confusing and didn't really add anything to the game.
This changes it so that the Equipment tab granted by either an outfitter
or a shipyard functions the same way regardless of whether a shipyard
is technically on the planet.

This also in tandem removes some empty shipyards throughout which
were there because of this odd mechanic, as they are no longer
necessary.