Made it so the Target Nearest key never targets escorts.
There was a half-baked measure implemented to do this, it seems, but
it didn't have any practical effect thanks to a special exception in
player_targetNearest. Excluding escorts from Target Nearest makes
sense because they're the ships you'd be least likely to need to
select with the keyboard, and them being candidates pretty much
entirely shuts off other options as a side-effect of them staying
close to you.