Finally added a number display for instant mode weapons.
I found a relatively simple API change that did just the trick: instead
of the "instant" parameter for weapset weapons showing a bool, it now
actually tells the first weapon set they're available in via instant
mode, much like the "weapset" parameter for actives. As a bonus, this
is an entirely compatible change; old code using it as a binary will
still work just as before.

Also made some tweaks to the Brushed GUI to make the new numbers more
seamless, which has a nice side-effect of making weapon displays more
consistent with active outfit displays.