Massively touched up the tutorial and escorts event, added follow tutorial.
This commit is getting way too large.

This started out as fixing up some things in the tutorial to make the
text a bit better, but then I went all-out and made a tutorial for
following as well. Then I noticed some deficiencies in the escort
handler event, so I fixed those, too. In particular, the escort handler
event will now report the speed of the ship you're hiring, so you can
more accurately decide if an escort is right for you.