Intro text is now always centered and narrow.
This changes a design where the image width was what the text size
was based on and the text was aligned to the left plus image with
if an image was there. That looked rather unprofessional. This
changes it so that it's always centered.

I'm not happy with how the images work, and I'm thinking of removing
them altogether, but decided to leave support for them in for now.