Reorganized things to eliminate a warning.
land_takeoff.h was being used to define a static variable for some
reason, rather than defining it in land.c and referencing it as an
extern in land.h. I don't know why, but I've changed it to the
latter, which is the more proper way to do something like this
anyway.