Started creating some new categories for outfit techs.
So, this is the paradigm: each ship class has its own collection of
each general type of outfits that it goes with, basic and elite
categories. The idea here is that planets will have a focus on
certain ship types (sometimes one, sometimes multiple), and on top
of that factions will have their own distinctions.