Made the Empire recruitment mission into a tutorial mission.
This forces the Empire recruitment mission to run at least once.
That could cause a problem during the tutorial, so I also added a
special condition to it requiring either the tutorial to be completed,
or the player to be out of the systems the tutorial uses (so the
tutorial missions can still be skipped).