Changed the mission marker colors to a gradient from yellow to red.
The new marker's color is blue, since for most colorblindness types
it's contrasting and it has a backup shape. For the "importance"
indication, it now goes from yellow to orange to red, with computer
missions having a noticeably brighter yellow. This seems to work
with all colorblindness types well enough (cone monochromats will
have the hardest time, as usual, but the difference between computer
and non-computer missions is still apparent and I ensured that, above
all, the markers have decent contrast so that they are visible).