Changed size colors.
This was a lot harder than the markers and I've literally worked all
night on this. The colors are the same saturation and value as the
colors that were used originally, but hue-shifted so that the sizes
are indicated by yellow, orange, and red, with blue used for indicating
a slot as required. I might want to adjust the exact values later.
Unfortunate I couldn't manage to accommodate wost-case-scenario
cone monochromacy because of that form of vision's effect of making
blue brighter, combined with the already difficult task of finding
a color that stands out enough from three different intensities for
a difference to be perceptible.