Renamed "Inca" to "Amaru" and removed "Maya".
The whole idea of naming planets after a name for a people is very,
very strange to me, not to mention, putting references to peoples
with such different cultures and histories as the Inca and the Maya
next to each other is so strange that it seems kinda racist to me.
I removed Maya (gameplay-wise it's redundant), and renamed "Inca" to
"Amaru", after Túpac Amaru II (who named himself after Túpac Amaru,
the last Sapa Inca of the Neo-Inca State).