Empire Recruitment now always sends to a planet with a mission computer.
The Empire Recruitment is now meant to serve as a conclusion to the
tutorial, transitioning the player into accepting jobs through the
mission computer to begin the game's main gameplay loop. This works
best if it sends the player to a planet which has a mission computer,
since it allows the player to immediately follow the suggestion to
go to the mission computer for ES missions rather than having to
explore first.