Increased the health of ships that are medium-heavy and heavier.
This is a part of some changes to the weapon balance. We'll be
increasing the damage of heavy weapons, and to compensate for that,
the health of the largest ships is being increased. This also has
the nice effect of stratifying health amounts more, which creates
more of an advantage for the largest ships.

Along the way, I had to convert some bioship outfits to the template
system which weren't already converted, and I noticed a small annoyance
with the way bio_template.py handled small number difference (fixed
by rounding to the nearest hundredths place).