Improved the closing-off of the combat practice mission.
There's a number of improvements here that basically amount to
actually ending the fight properly etc. One thing I had to do as a
part of one of them is fix an engine problem: previously, beams didn't
trigger the attacked hook except for the first frame they did damage,
as a side effect of a measure to avoid spamming the AI with messages
about being attacked. That's been fixed.