Added a pilot kill hook, cleaned up some pilot hook stuff.
At first I was just going to add a "kill" hook and be done with it,
but while fixing up the documentation, I also realized that the
behavior of the board and boarding hooks is absurd; both did exactly
the same thing (despite Naev's documentation claiming the board hook
was player-only, it triggered for all boarding instances just like
the boarding hook did), except the boarding hook had an extra parameter
that the board hook didn't.

In fact, that incorrect documentation claiming that the board hook
only applied to the player is responsible for some rather iffy code
that could malfunction if a pirate came along and boarded some
boarding target.

So what I did was change it so that the boarding hook does what its
naming suggest (hook a pilot boarding another pilot), while the board
hook does what the boarding hook used to do. Since the previous board
hook's implementation had a useless duplicate parameter, few Lua code
changes actually have to be made to accommodate this change, although
next I'll want to harden all those missions that use the board hook
and just assume the player did it.