Removed the standing cap system entirely.
I did some thinking about this and talked about it with my girlfriend
and I think this whole "faction standing cap" system needs to be
scrapped entirely. It's very confusing, and it hardlines the player
into getting reputation in only one particular way, which isn't a
very good gameplay experience. Actually, I think I remember us bringing
up how confusing and restrictive the rep cap system is years ago
with bobbens, but he said something along the lines of not wanting
reputation to be all about just killing ships.

Thing is, though, there's a much better way to achieve that: simply
reward more reputation for doing the story missions than for killing.
This is what MMOs seem to do.

The *only* "advantage" of the faction cap system I can think of is
that it forces the player to do the story, which is completely
antithetical to what I want this game to be. Sure, it might make
more sense for the Empire to only give you Empire ships after doing
their story, but this is quite an extreme measure, restricting the
players' freedom, for a rather insignificant benefit. I suppose this
also ties into bobbens' desire to force players to choose one or two
"major factions" to ally with and get the ships of, another design
idea I'm thinking is so controlling as to be anti-player.

All that considered, I'm taking the faction cap system out.