Bounties: Default to the faction that owns a system.
Some systems have higher presence flowing in from other systems
than that of the faction that technically owns them (such as Goddard).
This ensures that any owner of a system takes priority, and it only
switches over to others if the owner faction is not enemies with the
destroyed pilot.