Coupled together FLF and Frontier reputation.
This causes FLF rep to fall to no more than your Frontier rep if your
Frontier rep goes below zero; effectively, you cannot befriend the
FLF if you are enemies with the Frontier, and the Frontier will start
to hate you if you make enemies with the Frontier no matter how high
your FLF standing was. This makes sense lore-wise and will prevent
weirdness of someone who attacks the Frontier being hailed as a hero
of the Frontier.