Revert "Combined GUIs and maps into one "misc" folder with two sub-directories."
This reverts commit b29f3d9d56b2973841d9c55cfe29d85d885abfc3.

Turns out, the code is real fragile and actually specifically requires
maps to go in outfits/maps and nowhere else. There's a variable controlling
this, but there seems to be an implicit requirement that that variable
be under the outfits directory, too, I assume because parts of the
code depend on its own directory while other parts depend on it being
in outfits/.

This is an atrocious design, but for now, going back to the way it
was before. Maybe I'll fix the design some other time.