Buffed Gawain speed and takeoff speed.
The Gawain is not a combat ship at all, so let's allow it to have
tremendously more speed than anything else. Let's also buff that
takeoff speed buff; a minor reduction to land delay isn't a big deal,
but make it cut land delay to a fourth of the normal time, and
suddenly you can make four stops for missions that would normally
only allow time for one. (No real use-case for that yet, but I've
got some ideas for how to give it one.)