Added basic rescue code to the mining and combat tutorials.
This causes the player's ship to just be given laser cannons if they
have less than two weapons. It's not 100% perfect, but this should
at least make it so that it's very unlikely for a player to break
the mission.

I was originally going to add an additional fallback that bypasses
CPU requirements, but that could cause worse problems if they player
lands on a planet with no outfitter or shipyard.