Slightly improved how some options stuff is handled.
The only major effects of this change are that it falls back to
1280×720 instead of 800×600 when GetCurrentDisplayMode fails while
setting video defaults (not sure why it did a fallback to 800×600
to be honest), and I removed the insertion of the current resolution
as an option to the resolution list and instead ensured that the
default resolution is inserted. Everything else is just styling or
readability improvements.