Reorganized options in conf.h, changed default for doubletap afterburn.
The organization for options has historically been very inconsistent
and comments sometimes needlessly confusing. This is a start at
addressing that (though we'll want to reorganize the configuration
save and load stuff later on as well).

I reorganized everything in conf.h to align more closely with where
they appear in the in-game Options menu (if they do). While I was
doing that, I noticed that the doubletap afterburn option was handled
in a very kludgy manner and completely changed it.

Also changed double tap afterburn to disabled by default.