Added another DEBUG call, this time to check configured window size.
It seems the previous commit fixed the problem of the window size
textbox getting set to 640x480 due to it getting selected by default,
but I want to see if I can figure out why it's selecting that one
in the first place. After all, it should only choose that index if
it exactly equals the configured size.