This should fix the resolution list position bug in Windows.
Those debug prints showed what was going on! It seems under Windows
there's duplicate resolutions, which the code was aware of and
removed to avoid weirdness, but whoever wrote this forgot to compensate
for the removed entry when choosing what list index to select by default.
I've fixed that by counting duplicate resolutions and subtracting
duplicate resolutions from the resolution default setting.