Replaced "cycles" with shorter "galactic years".
The reason "cycles" wasn't replaced with "galactic years" in 0.4
was mainly because 50,000,000 seconds really isn't all that close
to a year. So I've changed it: we now have the galactic year, which
is 30,000,000 seconds (slightly less than an Earth year, which is
31,557,600 seconds). Yes, it means that you can't say that 2 cycles
is a clean 100,000,000 seconds, but come on, this isn't an actually
important consideration whatsoever. In the real world, if some new
time system was developed, its developers would bend over backwards
to make it as close to Earth time as possible (probably even closer
than this 30,000,000 seconds figure to be honest).

And much more important than theoretical clean numbers that the
player will never see is for players to be able to understand the
numbers we're throwing at them.