Completely gutted the whole player price knowledge system.
Things are still a bit of a mess, but now the prices reported aren't
affected by special "knowledge" requirements anymore (but the system
otherwise works the same as before). Variables storing this data have
not been completely removed, but saving and loading of the data, as
well as updating of the data, have been removed.