Added a custom variation of Naev's 7b9b93b.
I didn't check to confirm that the bug exists in Naikari, but given
the nature of it, there's a good chance that it does (since it seems
to have been caused simply by landing on a removed planet). The fix
I've implemented is a bit more complicated than Naev's due to some
noteworthy differences between the two, but the overall method is
the same.