OK, new paradigm for outfit groups.
The outfit groups I came up with before were way too specialized,
making it almost impossible to keep track of what I'm doing even
though I haven't even finished editing Empire space. A simplification
is needed, I think. This changes it to three tiers of the three
size types, which should be much more manageable. (Keeping the old
stuff in until it gets replaced, but moved it to the bottom so it's
out of the way when I'm editing.)