Fixed broken handling of maximized window and improved the method.
The previous code had something that was meant to tell the player
that the game couldn't resize itself if it was maximized… but this
isn't actually true. It *can* resize itself, it just wasn't bothering
to do so (by calling SDL_RestoreWindow). Getting rid of that and
just making it resize the window as the user asks avoids bugs which
included creating a faulty window if a change was actually attempted
and telling the player about not being able to resize the window
when that wasn't even attempted.