Changed the way autonav determines hostile presence.
This change makes it so that very distant enemies no longer cause
autonav to signify danger by resetting speed. Instead, speed is only
reset if a hostile ship is near to within range of either your weapons
or their weapons (with a more careful threshold if they either are
marked as hostile or are able to see you).