Completely changed the time system, removed DIY Nerds.
Part of a 2-step process to making units make more sense in Naikari.
For this part, we've completely changed the time system's in-universe
magnitudes, you could say. It works in fundamentally the same way,
but the second time unit in the time structure is the day, not the
hour, and in-universe time passes about 20 times faster than before
(650 in-universe seconds per second instead of 30 in-universe seconds
per second). The purpose of this is to allow us to use distance scales
which are more realistic while avoiding having ships go faster than
the speed of light. (The specific factor of 650 was calculated to make
the speed of 1300 approximately the speed of light, as the fastest
weapons in the game move at that speed.)

Next up: distance unit change.