Made missions check against the Empire Recruitment mission, not tut4.
I noticed that these missions were kind of clobbering the Empire
recruitment mission, which I don't want. I want the player to be
steered toward joining the Empire as a matter of course, so I want
to make sure there aren't distractions the first time you're offered
to join the Empire.