Manually set the kill reward for the pirate in tutorial4.
The auto-generated kill reward is semi-random, but somewhere in the
region of 100 k¢, which is a bit excessive considering the fact that
it's a heavily weakened Hyena (and normally should be the player's
first kill). 20 k¢ seems more appropriate; enough to communicate to
the player that killing pirates is worth it, but not so much that
it gives the player a massive easy boost (and static so that the
player doesn't benefit or hurt from the randomness).