Removed the landing service from Eiroik.
I don't like the way Naev's Collective campaign uses this planet,
and will not be duplicating it. With that in mind, there's no need
to allow landing here.

Getting rid of Eiroik as a landing point will also enable getting
rid of that exception to auto-saving bobbens put in as a band-aid
solution to planets like this being able to cause broken saves.