Made the refuel failure messages a bit more flexible.
This makes it possible to override the messages for when payment
is received and for when a pilot cannot refuel you, and also changes
the message for busy pilots to one that reads like a system message.
The purpose of this is so I can implement refueling behavior with the
Collective drones that makes sense.