Don't allow AI ships to land on planets without the land service.
This is a bit of an edge case, but the code only actually checked
if a planet was inhabited, not if it had the land service,
specifically. These don't mean the same thing. "Inhabited" is a
virtual "service" indicating that the planet is owned by a faction
(rather than being factionless) and has at least one service. While
the editor doesn't allow a planet to have other services without
the land service (and thus it normally isn't possible to have an
inhabited planet without the land service; indeed, that case would
be in error), the code does not actually enforce this.