Added a "fuel request" event.
This event causes an NPC to spawn which is out of fuel and needs
assistance (similar to how the player can need to buy fuel). The
chance of it happening is low so that it doesn't become obnoxious
(meaning there's a good chance of a player just never running into
it in a playthrough, but that's fine).