Added a feature to make the pilot face function have perfect accuracy.
The new feature is a "direction destination" which can be passed to
the physics system. When this is set to something other than -1, the
code which turns an object for directional velocity will stop if the
rotation brings the object to that direction, and then it will set
the rotation to that exact direction. This fixes a problem where ships
with incredibly fast rotation (like an Empire Shark with Zephyr Engine
and two steering thrusters) could get stuck trying to rotate between
two points that were not quite the direction they were trying to
turn in, especially when time compression was active.

I suspect a similar problem from delta timing is the cause for the
glitch that happens at jump points at times that causes a ship with
very high acceleration to get stuck spinning around for awhile.
Implementing this same "velocity destination" system for velocity
is much more complicated given that it's two-dimensional. Will have
to think on this.